prooflayer/requirements.txt
fastapi
uvicorn
python-multipart
pillow
opencv-python-headless
numpy
pycryptodome
exifread
pymupdf
pdfminer.six
